Add: zip() to replace zipMap() and enable other usages
Before we had lazily computed arrays, implementing zip would have required copying the contents of both arrays into a new DenseArray which would then be promptly thrown away. zipMap() was a way of avoiding that, but it can now be dispensed with.

We also expending some generic function's type signatures to be a little more general.